backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-docx
psycopg2-binary
python-dotenv
